move qcut into transforms
