+++ tests/qcut/test_qcut.py
@@ -29,7 +29,7 @@ from scipy.stats import unitary_group
 
 import pennylane as qml
 from pennylane import numpy as np
-from pennylane.qcut import qcut
+from pennylane.transforms.qcut import qcut
 from pennylane.wires import Wires
 
 pytestmark = pytest.mark.qcut
